Add files via upload
This program demonstrates 2 disciplines how AI models are trained and using a BST as compared to using an Array. Two versions of the same app exist, they do exactly the same thing, but they differ in the fact that the other is using an array, and the other one is using a Binary Search Tree. There is a textfile GenericsKB that contains knowledge in the format of <term<tab>knowledge<tab>confidence rate> containing about 50 000 lines. You can add knowledge and search for existing knowledge and much more do it is quite interesting look at it.